requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
python-docx
PyPDF2
docling
google-generativeai
sentence-transformers
numpy
scikit-learn
streamlit
gradio
motor
pymongo
pytest
pytest-asyncio
requests
beautifulsoup4
networkx
openai 